+++ GetData/GetDataYML.cs
@@ -175,7 +175,7 @@ namespace wackydatabase.GetData
 
         private StatusData GetStatusData(StatusEffect effect)
         {
-
+            
             //effect.m_icon = effect.m_icon.name;
             var f2 = effect.GetType();
             WMRecipeCust.WLog.LogInfo("             StatusEffect " + effect.name);
@@ -185,7 +185,7 @@ namespace wackydatabase.GetData
             { // the Skill levelup could work LevelUpSkillOnBreak
                 shield.AbsorbDmg = Functions.getCast<float>(f2, "m_absorbDamage", effect);
                 shield.AbsorbDmgWorldLevel = Functions.getCast<float>(f2, "m_absorbDamageWorldLevel", effect);
-                shield.LevelUpSkillFactor = Functions.getCast<int>(f2, "m_levelUpSkillFactor", effect);
+                shield.LevelUpSkillFactor = Functions.getCast<float>(f2, "m_levelUpSkillFactor", effect);
                 shield.TtlPerItemLevel = Functions.getCast<int>(f2, "m_ttlPerItemLevel", effect);
                 shield.AbsorbDmgPerSkill = Functions.getCast<float>(f2, "m_absorbDamagePerSkillLevel", effect);
